requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
boost=1.72.0=py37h96a391d_0
boost-cpp=1.72.0=h0caebb8_0
bzip2=1.0.8=hfa6e2cd_2
ca-certificates=2020.12.5=h5b45459_0
cairo=1.16.0=h8a5e50b_1003
certifi=2020.12.5=py37h03978a9_1
cffi=1.14.0=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.1=py_0
colorama=0.4.3=py_0
cryptography=2.8=py37h7a1dbc1_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
dm-sonnet=1.23=pypi_0
fcd=1.1=pypi_0
fcd-torch=1.0.7=pypi_0
flask=1.1.1=py_1
flask-cors=3.0.8=py_0
freetype=2.10.1=ha9979f8_0
gast=0.3.3=py_0
gettext=0.19.8.1=hb01d8f6_1002
glib=2.58.3=py37hb6b7ca9_1004
google-pasta=0.2.0=pypi_0
grpcio=1.27.2=py37h351948d_0
h5py=2.8.0=py37hf7173ca_0
hdf5=1.8.18=vc14h7a021fe_0
icc_rt=2019.0.0=h0cc432a_1
icu=64.2=he025d50_1
idna=2.9=py_1
importlib-metadata=1.6.0=py37hc8dfbb8_0
importlib_metadata=1.6.0=0
intel-openmp=2020.0=166
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
itsdangerous=1.1.0=py37_0
jedi=0.16.0=py37_1
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9c=hfa6e2cd_1001
jsonschema=3.2.0=py37hc8dfbb8_1
jupyter_core=4.6.3=py37hc8dfbb8_1
kaleido-core=0.2.1=h8ffe710_0
keras=2.3.1=py37h21ff451_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37heaa310e_0
libclang=9.0.1=default_hf44288c_0
libffi=3.2.1=h6538335_1007
libgpuarray=0.7.6=hfa6e2cd_1003
libiconv=1.15=hfa6e2cd_1006
libpng=1.6.37=hfe6a214_1
libprotobuf=3.11.4=h7bd577a_0
libtiff=4.1.0=h885aae3_6
lz4-c=1.9.2=h33f27b4_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mako=1.1.0=py_0
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
mathjax=2.7.7=0
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py37h911224e_0
mkl=2019.5=281
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
molsets=1.0=pypi_0
msys2-conda-epoch=20160418=1
nbformat=5.0.4=py_0
networkx=2.6.3=pypi_0
numpy=1.18.1=py37h93ca92e_0
numpy-base=1.18.1=py37hc3f5095_1
olefile=0.46=py_0
openssl=1.1.1k=h8ffe710_0
pandas=1.0.3=py37h9b59f54_0
parso=0.6.2=py_0
pcre=8.44=h6538335_0
pickleshare=0.7.5=py37_0
pillow=7.1.1=py37h91e7a8d_0
pip=20.0.2=py37_1
pixman=0.38.0=hfa6e2cd_1003
plotly=4.6.0=py_0
plotly-orca=1.3.1=1
pomegranate=0.12.0=pypi_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
protobuf=3.11.4=py37h33f27b4_0
psutil=5.7.0=py37he774522_0
pycairo=1.19.1=py37h6430cfb_3
pycparser=2.20=py_0
pygments=2.6.1=py_0
pygpu=0.7.6=py37hc8d92b1_1000
pyopenssl=19.1.0=py37_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h6538335_1
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.16.0=py37h8055547_0
pysocks=1.7.1=py37_0
python=3.7.7=h60c2a47_0_cpython
python-dateutil=2.8.1=py_0
python-kaleido=0.2.1=pyhd8ed1ab_0
python_abi=3.7=1_cp37m
pytz=2019.3=py_0
pywin32=227=py37hfa6e2cd_0
pyyaml=5.3.1=py37h8055547_0
qt=5.12.5=h7ef1ec2_0
rdkit=2020.03.1=py37hea70aac_1
requests=2.23.0=py37_0
retrying=1.3.3=py37_2
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.4.1=py37h9439919_0
seaborn=0.10.0=py_0
setuptools=46.1.3=py37_0
six=1.14.0=py37_0
sqlite=3.31.1=he774522_0
tensorboard=1.14.0=py37he3c9ec2_0
tensorflow=1.14.0=gpu_py37h5512b17_0
tensorflow-base=1.14.0=gpu_py37h55fc52a_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py37_1
theano=1.0.4=py37h6538335_1001
tk=8.6.10=hfa6e2cd_0
torch=1.10.0=pypi_0
tornado=6.0.4=py37hfa6e2cd_0
tqdm=4.62.3=pypi_0
traitlets=4.3.3=py37_0
typing-extensions=3.10.0.2=pypi_0
urllib3=1.25.8=py37_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
vs2015_win-64=14.0.25420=h55c1224_11
wcwidth=0.1.9=py_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h2fa13f4_0
yaml=0.2.3=he774522_0
zipp=3.1.0=py_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.4=h9f78265_3
